CLIENT: Need to do changes in InputContainer, but for now I received the receiver user from the goToPrivateChat slice, and tried to get that receiver id using the useGetUserIdMutation. Cleared almost all the MessageContainer, changed the description prop in NewChatUser to be optional, in the NewChatUserList I got the data of the registered users using the useGetContactsQuery and showed the existing users in the modal, chandged the Person interface in PeopleContainer, NewChatUserList, userListSlice to PrismaUser interdace, stopped the use of usersSlice in the chat modal, in the Person interface changed the id prop to email prop. Created a new interface PrismaUser, to match the User model of the prisma schema, which is the output of the users queries and mutations in the client components, created a Message interface, because two of the props of the PrismaUser interface is an array of messages, and it also matches the Message model in the prisma schema. Created getContactsApi, which will get all the contacts (users) that are registered to the application.
SERVER: In prisma schema, changed the Message column to a column in User model of received messages and sent messages. Did the same in Message model, and called the columns senderId and receiverId.
In authService, signupLocal and loginEmailLocal routes, I selected which fields of the user to include and which to not, the return from the routed will be the user as well as the tokens, in the message controller I added a otherUserId parameter in the getMessages route, in the message service I changed the logic of the getMessages route, so it will get receivedMessages and sentMessages between the current user and otherUserId with the help of prisma, and will return them.
Added a receiverId field in the messageDto, added getContacts route and getAllUsers route (maybe the last one is not necessary) in the user controller, in the user service added a getContacts logic and getAllUsers logic.